Week 1 Practice Project
Unfinished. I wanted to make it so that if someone entered a zip code they'd get the name of the city associated with it back. So far all it does is take the zip and city and combine them together. I think it only takes the first of the array. How would I make it so that it displays all from the array?